171004 (수) 어제 한 것 - 굿 씨앤티 누르면 ajax 로 처리되며 굿 씨앤티 테이블에 userid 추가, 삭제. - 굿 버튼 눌렀을 때 아이디가 있는지 없는지 확인 // 아이디 있으면 + 1, 없으면 -1 - 굿 버튼 누르면 + 1, 다시 누르면 -1 기능 구현! - readPage(상세페이지) g/b 카운트 제대로 나옴 - 리스트의 g/b, 상세페이지의 g/b 의 넘버 포맷팅 완료
오늘 할 것
- good, bad 다른 아이디도 중복없이 등록 가능 // spam까지 동일하게 구현!

예상 되는 오류, 버그
-

도움이 필요한 것
-

+++ src/main/java/org/zerock/domain/BoardVO.java
@@ -20,7 +20,7 @@ public class BoardVO {
 
     private int goodcnt;
     private int badcnt;
-    private int gbcnt;
+    private double gbcnt;
     private int spamcnt;
 
     private int viewcnt;
@@ -197,11 +197,11 @@ public class BoardVO {
         this.badcnt = badcnt;
     }
 
-    public int getGbcnt() {
+    public double getGbcnt() {
         return gbcnt;
     }
 
-    public void setGbcnt(int gbcnt) {
+    public void setGbcnt(double gbcnt) {
         this.gbcnt = gbcnt;
     }
